ODIN-121 #comment Initial Checkin for nowait submit, and status query. ODIN-121 #time 10h
Update commons-exec to 1.3 during investigations as to why we don't get output
from the PumpStreamHandler (it's likely that dfuplus doesn't flush stdout
when it's killed).
Added nowait=1 to spray/despray, which means we now execute multiple
action=status wuid=X commands and wait for the status to change.
Removed SimpleOutputParser which is now split into a JobSubmitParser and
a JobStatusParser; they now extend LogOutputStream so that they can process
each logging line as it is flushed by dfuplus (still not guaranteed I guess,
because we still don't get it if we watchdog-kill it).
- config break : Removed timeout (it is now monitor-interval)
- config break : Removed parse-output (we always parse output).