INTERLOK-1660 Added a max-wait INTERLOK-1660 #time 2h
- Switch to using a Future<?> rather than a loop in execute (it's still
really the same IMO); but we can cancel the future during stop.
- Cancel the future during stop(); in a StandardWorkflow that won't
actually fire, until the message is processed; in a PoolingWorkflow stop
is fired, if we exceed the pool shutdown wait.
- Made the default max-wait to be an hour if null; toyed with making it
infinite in the absence of configuration that but seems deliberately
obtuse as we don't want a stuck service you can never stop.